fix(diff): resolve base files to snapshots before mapping; escape ref labels via shared util
Addresses review feedback on #3880:

P1 — Move snapshot resolution out of computeOutputDiffStats into
OutputNode.exec so the precomputed mapping is built against snapshot
locations. Previously OutputNode built the mapping with live baseFiles
and passed it in; computeOutputDiffStats then resolved snapshots
internally but the precomputedMapping shortcut bypassed resolution,
so in-place workflow diffs still collapsed to 0/0.

Now OutputNode resolves once via resolveBaseFilesForDiff and threads
the resolved baseFiles through traceFileLineage, handleLatexdiff,
summarizeRound, and getRoundOutput. computeOutputDiffStats no longer
knows about executionIds; its docstring documents the caller's
responsibility to pre-resolve.

P2 — Escape LaTeX reference labels before interpolating into
data-label attributes and inner text. UserMessage now renders
structured-delivery content through processMarkdownContent, so a
crafted label in model output could otherwise break out of the
attribute and inject DOM. Extracted escapeAttr/escapeText from
src/tools/subagentResults.ts to @shared/utils/xmlEscape so both
agent-core (XML emission) and frontend (HTML interpolation) share
one source of truth. subagentResults re-exports for compat.
