+++ src/agent/output/diffComputation.ts
@@ -9,15 +9,10 @@ import { diff_match_patch } from 'diff-match-patch';
 
 import * as logger from '@agent/core/logger';
 import type { DiffStats } from '@agent/types/DiffTypes';
-import type {
-  ExecutionId,
-  OutputFileInfo,
-  FileLocation,
-} from '@shared/schemas';
+import type { OutputFileInfo, FileLocation } from '@shared/schemas';
 import { flexibleFS, getComparablePath } from '@utils/files';
 import { countLines } from '@utils/text/stringUtils';
 
-import { resolveBaseFilesForDiff } from './snapshotResolution';
 import { traceFileLineage } from './lineageMapping';
 import { ensureRound, type OutputState } from './outputState';
 import type { RoundFileMapping } from './types';
@@ -69,24 +64,24 @@ async function computeDiffStats(
 // Public API
 // ============================================================================
 
-/** Computes diff stats for all output files in a round. */
+/** Computes diff stats for all output files in a round.
+ *
+ *  Callers in workflows must pass snapshot-resolved baseFiles (via
+ *  resolveBaseFilesForDiff) when an executionId is available; passing
+ *  the live workspace path would collapse in-place diffs to 0/0. The
+ *  precomputedMapping, if provided, must have been built against the
+ *  same snapshot-resolved baseFiles — otherwise the mapping's base
+ *  locations still point at the overwritten files. */
 export async function computeOutputDiffStats(
   state: OutputState,
   baseFiles: FileLocation[],
   currRound: number,
   precomputedMapping?: RoundFileMapping,
-  options?: { isRewrite?: boolean; executionId?: ExecutionId },
+  options?: { isRewrite?: boolean },
 ): Promise<OutputFileInfo[]> {
   const roundOutputs = ensureRound(state, currRound);
-  // Substitute workspace base files with their pre-run snapshot when one
-  // exists. In-place workflows overwrite the live workspace file before
-  // we get here; reading it as the diff base would collapse to 0/0.
-  const diffBaseFiles = await resolveBaseFilesForDiff(
-    baseFiles,
-    options?.executionId,
-  );
   const mapping =
-    precomputedMapping ?? traceFileLineage(state, diffBaseFiles, currRound);
+    precomputedMapping ?? traceFileLineage(state, baseFiles, currRound);
   const suppressLineage = options?.isRewrite === false;
 
   return Promise.all(
@@ -111,12 +106,8 @@ export async function computeOutputDiffStats(
       // "methods") don't match the base filename via basename strategies.
       // If no diff base was found but there is exactly one base file, use it
       // so the diff stats reflect real changes against the original.
-      if (
-        !diffBaseLocation &&
-        !originalLocation &&
-        diffBaseFiles.length === 1
-      ) {
-        const candidate = diffBaseFiles[0];
+      if (!diffBaseLocation && !originalLocation && baseFiles.length === 1) {
+        const candidate = baseFiles[0];
         if (getComparablePath(candidate) !== locationPath) {
           diffBaseLocation = candidate;
         }
