+++ src/agent/output/roundSummary.ts
@@ -44,6 +44,11 @@ export async function summarizeRound(
     stage?: AgentLogStage;
     mapping?: RoundFileMapping;
     isRewrite?: boolean;
+    /** Snapshot-resolved base files. Required for in-place workflows so
+     *  diff stats are computed against the pre-run snapshot rather than
+     *  the overwritten workspace file. When omitted, falls back to
+     *  deps.baseFiles (live workspace paths). */
+    baseFiles?: FileLocation[];
   },
 ): Promise<RoundSummary> {
   return withOutputStage(
@@ -56,10 +61,10 @@ export async function summarizeRound(
 
       const fileInfos = await computeOutputDiffStats(
         state,
-        deps.baseFiles,
+        options.baseFiles ?? deps.baseFiles,
         currRound,
         options.mapping,
-        { isRewrite: options.isRewrite, executionId: deps.executionId },
+        { isRewrite: options.isRewrite },
       );
       data.outputs = fileInfos;
       const storageKey = getStorageKey(state);
@@ -96,7 +101,7 @@ export async function getRoundOutput(
   state: OutputState,
   baseFiles: FileLocation[],
   round: number,
-  options?: { isRewrite?: boolean; executionId?: string },
+  options?: { isRewrite?: boolean },
 ): Promise<RoundOutput> {
   const data = ensureRoundData(state, round);
